Make app run
Edit the activities and make application run for the first time. Now you can see top 20 images using RecyclerView in MainActivity and if you click on one item the app navigates you to the GalleryActivity to see full details about the image. Also a button opens the Google Maps service